色の設定のJSONファイルを.colterに作成してくれる
